requirements.txt
Axiom==0.6.0
BeautifulSoup==3.2.1
Brlapi==0.5.7
Coherence==0.6.6.2
Django==1.5.1
Epsilon==0.6.0
GnuPGInterface==0.3.2
Louie==1.1
Mako==0.7.0
MarkupSafe==0.15
Nevow==0.10.0
PAM==0.4.2
PIL==1.1.7
PyOpenGL==3.0.1
SOAPpy==0.12.0
South==0.8.1
Twisted-Conch==12.0.0
Twisted-Core==12.0.0
Twisted-Web==12.0.0
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
configobj==4.7.2
cups==1.0
defer==1.0.6
django-oauth-plus==2.1.2
django-tastypie==0.10.0
feedparser==5.1.2
fpconst==0.7.2
gdata==2.0.17
gnome-app-install==0.4.7-nmu1
httplib2==0.8
lazr.restfulclient==0.12.0
lazr.uri==1.0.3
louis==2.4.1
mechanize==0.2.5
numpy==1.6.2
oauth==1.0.1
oauth2==1.5.211
pexpect==2.4
pyOpenSSL==0.13
pyasn1==0.1.3
pycrypto==2.6
pycurl==7.19.0
pyserial==2.5
pysqlite==2.6.3
python-apt==0.8.8.2
python-dateutil==2.1
python-debian==0.1.21
python-debianbts==1.11
python-mimeparse==0.1.4
pyxdg==0.19
rdflib==2.4.2
reportbug==6.4.4
reportlab==2.5
simplejson==2.5.2
six==1.4.1
tagpy==0.94.8
uTidylib==0.2
unattended-upgrades==0.1
vboxapi==1.0
virtualenv==1.10.1
wadllib==1.3.0
wsgiref==0.1.2
zope.interface==3.6.1
